Room Occupancy Monitor Code
This code checks for the amount of users that enter and exit a room, keeping track of a count. The user is allowed to set their own maximum room occupancy (Which maxes out at 100 --> Due to class requiring that being max), and once that capacity is breached, the buzzer will go off, making a loud noise to alert any other people that the room is at maximum capacity, whilst also spinning a motor which alerts users additionally using a flag. There are two sensors (both Infared), one of which is for users entering and exiting, doing the necessary reaction upon each condition being fulfilled. The code also displays the count to the user.